WIP - Sketch out a Guides controller
In order to keep our documentation in sync, we wanto publish it as
part of our regular releases and present it on the Convene website.

I'm considering setting it up so that we can pull docs in (either from
Google Docs, or maybe from the Feature definitions... TBD) but I need to
think a bit more about how we may want to do that.